move classes to common
