+++ backend/common/src/main/java/ai/verta/modeldb/common/exceptions/UnavailableException.java
@@ -1,5 +1,6 @@
-package ai.verta.modeldb.exceptions;
+package ai.verta.modeldb.common.exceptions;
 
+import ai.verta.modeldb.exceptions.ModelDBException;
 import com.google.rpc.Code;
 
 public class UnavailableException extends ModelDBException {
