+++ backend/src/main/java/ai/verta/modeldb/entities/config/HyperparameterElementMappingEntity.java
@@ -1,7 +1,7 @@
 package ai.verta.modeldb.entities.config;
 
 import ai.verta.modeldb.entities.ExperimentRunEntity;
-import ai.verta.modeldb.exceptions.InternalErrorException;
+import ai.verta.modeldb.common.exceptions.InternalErrorException;
 import ai.verta.modeldb.exceptions.ModelDBException;
 import ai.verta.modeldb.versioning.HyperparameterValuesConfigBlob;
 import io.grpc.Status;
