+++ backend/src/main/java/ai/verta/modeldb/experiment/ExperimentDAORdbImpl.java
@@ -8,6 +8,7 @@ import ai.verta.modeldb.*;
 import ai.verta.modeldb.authservice.RoleService;
 import ai.verta.modeldb.common.authservice.AuthService;
 import ai.verta.modeldb.common.collaborator.CollaboratorUser;
+import ai.verta.modeldb.common.exceptions.NotFoundException;
 import ai.verta.modeldb.dto.ExperimentPaginationDTO;
 import ai.verta.modeldb.entities.*;
 import ai.verta.modeldb.exceptions.*;
