+++ backend/src/main/java/ai/verta/modeldb/utils/ModelDBHibernateUtil.java
@@ -27,7 +27,7 @@ import ai.verta.modeldb.entities.metadata.MetadataPropertyMappingEntity;
 import ai.verta.modeldb.entities.versioning.*;
 import ai.verta.modeldb.exceptions.AlreadyExistsException;
 import ai.verta.modeldb.exceptions.ModelDBException;
-import ai.verta.modeldb.exceptions.UnavailableException;
+import ai.verta.modeldb.common.exceptions.UnavailableException;
 import com.google.common.base.Joiner;
 import io.grpc.health.v1.HealthCheckResponse;
 import java.sql.*;
